+++ NexTube.Application/CQRS/Identity/Users/Commands/GetChannelInfo/GetChannelInfoCommandHandler.cs
@@ -7,7 +7,7 @@ using NexTube.Application.Common.DbContexts;
 using NexTube.Domain.Entities;
 
 
-namespace NexTube.Application.CQRS.Identity.Users.Commands.GetUser
+namespace NexTube.Application.CQRS.Identity.Users.Commands.GetChannelInfo
 {
     public class GetUserCommnadHandler : IRequestHandler<GetChannelInfoCommand, GetChannelInfoCommandResult>
     {
